Refactored TestTransactionFrame to own file and small fixes
