+++ src/test/TestAccount.h
@@ -6,6 +6,7 @@
 
 #include "crypto/SecretKey.h"
 #include "transactions/TransactionFrame.h"
+#include "transactions/test/TransactionTestFrame.h"
 #include "xdr/Stellar-ledger-entries.h"
 #include "xdr/Stellar-transaction.h"
